Return proper Node subclasses for comment, pi and text nodes (#37)
This is a breaking change.

Previously, libxmljs2 always returned an Element for comment nodes, text
nodes and processing instruction nodes. Change the code such that we now
return the proper subclasses of Node (Comment, Text and
ProcessingInstruction) instead of wrapping them in an Element.

Now code like

    const doc = libxml.parseXml('<root>text</root>')
    const text = doc.child(0)
    console.log(text instanceof libxml.Text) // true

will work as expected.